# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Citycard",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cities/{cityId}/transport": {
            "get": {
                "tags": [
                    "Transport"
                ],
                "summary": "Get all city's transport by cityId",
                "operationId": "e232e90c2e4752ecedfe2bba86b1f8bc",
                "parameters": [
                    {
                        "name": "cityId",
                        "in": "path",
                        "description": "ID of city",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "city_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "route_number": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "transport_type": {
                                                        "type": "string",
                                                        "example": "bus"
                                                    },
                                                    "route_endpoint_1": {
                                                        "type": "string",
                                                        "example": "Північна"
                                                    },
                                                    "route_endpoint_2": {
                                                        "type": "string",
                                                        "example": "Південна"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transport"
                ],
                "summary": "Create a new city's transport route by cityId",
                "operationId": "71090e962a9b645fa9d5e52acaa936a5",
                "parameters": [
                    {
                        "name": "cityId",
                        "in": "path",
                        "description": "ID of city",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "route_number": {
                                                "type": "number"
                                            },
                                            "transport_type": {
                                                "type": "string"
                                            },
                                            "route_endpoint_1": {
                                                "type": "string"
                                            },
                                            "route_endpoint_2": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "route_number": "1",
                                    "transport_type": "Автобус",
                                    "route_endpoint_1": "Південна"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "city_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "route_number": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "transport_type": {
                                            "type": "string",
                                            "example": "Автобус"
                                        },
                                        "route_endpoint_1": {
                                            "type": "string",
                                            "example": "Північна"
                                        },
                                        "route_endpoint_2": {
                                            "type": "string",
                                            "example": "Південна"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The route_number field format is invalid."
                                        },
                                        "error": {
                                            "properties": {
                                                "route_number": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The route_number field format is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}